backend/requirements.txt
Django>=5.2.4
djangorestframework>=3.15.2
djangorestframework-simplejwt>=5.3.1
django-cors-headers>=4.5.0
django-filter>=24.3
dj-database-url>=2.2.0
psycopg[binary]>=3.2.3
boto3>=1.35.0
python-dotenv>=1.0.1